Move migration base exception to a separate exception module
